requirements.txt
djangorestframework
django-filter  # Filtering support
django-cors-headers